Turn previous integration tests unit tests.
